add test for validating only one of RunConfiguration or RunName is present
